+++ packages/core/di/factory-creator.ts
@@ -1,5 +1,7 @@
+import 'reflect-metadata';
 import { TProviders, TInjectTokenProvider, TUseClassProvider, TUseValueProvider } from '../types';
 import { Injector } from './injector';
+import { metadataOfInjectable, metadataOfOptional, metadataOfHost, metadataOfSelf, metadataOfSkipSelf } from './metadata';
 
 /**
  * use injector to create arguments for constructor
@@ -13,8 +15,8 @@ export function injectionCreator(_constructor: Function, injector?: Injector): a
     const args: any[] = [];
 
     let _needInjectedClass: any[] = [];
-    if ((_constructor as any)._needInjectedClass) _needInjectedClass = (_constructor as any)._needInjectedClass;
     if ((_constructor as any).injectTokens) _needInjectedClass = (_constructor as any).injectTokens;
+    else _needInjectedClass = Reflect.getMetadata(metadataOfInjectable, _constructor) || [];
 
     // build $privateProviders into injector of component
     if ((_constructor.prototype as any).$privateProviders as TProviders) {
@@ -29,37 +31,60 @@ export function injectionCreator(_constructor: Function, injector?: Injector): a
         }
     }
 
+    // 干预等级 @SkipSelf > @Self > @Host > @Optional
+    const skipSelfList: number[] = Reflect.getMetadata(metadataOfSkipSelf, _constructor) || [];
+    const selfList: number[] = Reflect.getMetadata(metadataOfSelf, _constructor) || [];
+    const hostList: number[] = Reflect.getMetadata(metadataOfHost, _constructor) || [];
+    const optionalList: number[] = Reflect.getMetadata(metadataOfOptional, _constructor) || [];
+
+    // find instance from provider
     const needInjectedClassLength = _needInjectedClass.length;
     for (let i = 0; i < needInjectedClassLength; i++) {
+        // 构建冒泡层数
+        let bubblingLayer: number | 'always' = 'always'; 
+        if (hostList.indexOf(i) !== -1) bubblingLayer = 1;
+        if (selfList.indexOf(i) !== -1) bubblingLayer = 0;
+
+        // 构建冒泡开始的injector
+        let findInjector = injector;
+        if (skipSelfList.indexOf(i) !== -1) findInjector = injector.parentInjector;
+
         const key = _needInjectedClass[i];
-        if (injector) {
-            if (injector.getInstance(key)) {
-                args.push(injector.getInstance(key));
+
+        if (findInjector.getInstance(key, bubblingLayer)) {
+            args.push(findInjector.getInstance(key, bubblingLayer));
+            continue;
+        } else {
+            // use @Optional will return null
+            if (optionalList.indexOf(i) !== -1) {
+                args.push(null);
                 continue;
-            } else {
-                let findService = injector.getProvider(key);
+            }
 
-                if (findService) {
-                    if (!findService.useClass && !findService.useValue) findService = findService;
-                    if (findService.useClass) findService = findService.useClass;
-                    if (findService.useValue) {
-                        args.push(findService.useValue);
-                        continue;
-                    }
+            let findService = findInjector.getProvider(key, bubblingLayer);
 
-                    const serviceParentInjector = injector.getParentInjectorOfProvider(findService);
-                    const serviceInjector = serviceParentInjector.fork();
-
-                    if (findService.isSingletonMode === false) {
-                        args.push(factoryCreator(findService, serviceInjector));
-                        continue;
-                    } else {
-                        const serviceInStance = factoryCreator(findService, serviceInjector);
-                        injector.setInstance(key, serviceInStance);
-                        args.push(serviceInStance);
-                        continue;
-                    }
+            if (findService) {
+                if (!findService.useClass && !findService.useValue) findService = findService;
+                if (findService.useClass) findService = findService.useClass;
+                if (findService.useValue) {
+                    args.push(findService.useValue);
+                    continue;
                 }
+
+                const serviceParentInjector = findInjector.getParentInjectorOfProvider(findService, bubblingLayer);
+                const serviceInjector = serviceParentInjector.fork();
+
+                if (findService.isSingletonMode === false) {
+                    args.push(factoryCreator(findService, serviceInjector));
+                    continue;
+                } else {
+                    const serviceInStance = factoryCreator(findService, serviceInjector);
+                    findInjector.setInstance(key, serviceInStance);
+                    args.push(serviceInStance);
+                    continue;
+                }
+            } else {
+                throw new Error(`In injector could'nt find ${key}`);
             }
         }
     }