+++ packages/core/di/index.ts
@@ -1,4 +1,10 @@
+export * from './metadata';
 export * from './injectable';
 export * from './factory-creator';
 export * from './injector';
 export * from './injected';
+export * from './self';
+export * from './skip-self';
+export * from './optional';
+export * from './host';
+// export * from './inject';