+++ packages/core/di/metadata.ts
@@ -1,0 +1,9 @@
+export const metadataOfInjectable = 'nvMetadataOfInjectable';
+
+export const metadataOfOptional = 'nvMetadataOfOptional';
+
+export const metadataOfHost = 'nvMetadataOfHost';
+
+export const metadataOfSelf = 'nvMetadataOfSelf';
+
+export const metadataOfSkipSelf = 'nvMetadataOfSkipSelf';